Main script can now update both Dockerfiles and Concourse templates
- Rename update_script.rb to dependabot-main.rb
- dependabot-main.rb can now be used to update both Dockerfiles and Concourse templates, by setting the FEATURE_PACKAGE environmental variable to either docker or concourse
- .rubocop.yml is removed in favor of the upstream one at the project path
- modify the Dockerfile in preparation for moving it to linux-image-build repository
