--- docker/.rubocop.yml
@@ -1,23 +1,0 @@
-# Layout
-Layout/LineLength:
-  Max: 110
-
-# Style
-Style/CommentedKeyword:
-  Enabled: false
-Style/Documentation:
-  Enabled: false
-
-# Metrics
-Metrics/ClassLength:
-  Enabled: false
-Metrics/MethodLength:
-  Enabled: false
-Metrics/AbcSize:
-  Enabled: false
-Metrics/CyclomaticComplexity:
-  Enabled: false
-Metrics/PerceivedComplexity:
-  Enabled: false
-Metrics/BlockLength:
-  Max: 1000
