+++ docker/Dockerfile
@@ -1,17 +1,39 @@
-FROM docker.ci.pix4d.com/dependabot-ruby-base:latest AS main
-LABEL "maintainer"="integration_team@pix4d.com"
-LABEL "description"="Inspect Concourse pipelines and find possible updates to Docker images \
-    checking a specified registry."
+FROM ruby:2.7-alpine AS build
 
-COPY concourse /home/dependabot/concourse
-COPY Gemfile /home/dependabot/
-COPY update_script.rb /home/dependabot/
+WORKDIR /home/dependabot/
+
+ENV BUNDLE_PATH="/home/dependabot/.bundle" \
+  BUNDLE_BIN=".bundle/binstubs" \
+  PATH=".bundle/binstubs:$PATH:/home/dependabot/.bundle/bin"
+
+COPY common /home/dependabot/common
+COPY docker /home/dependabot/docker
+
+RUN apk update && \
+  apk add --no-cache \
+  g++ \
+  make && \
+  gem update --system && \
+  gem install bundler --no-document
+
+WORKDIR /home/dependabot/docker/
 
 RUN bundle install
 
-FROM main AS tests
+FROM build AS tests
 
-WORKDIR /home/dependabot/concourse/
+WORKDIR /home/dependabot/docker/
 RUN bundle exec rspec ./spec/
 
-FROM main
+FROM ruby:2.7-alpine AS main
+LABEL "maintainer"="platform_ci_team@pix4d.com"
+LABEL "description"="Inspect Dockerfiles and Concourse pipelines to find possible updates to Docker images \
+  by checking a specified registry."
+
+WORKDIR /home/dependabot/
+
+ENV BUNDLE_PATH="/home/dependabot/.bundle" \
+  BUNDLE_BIN=".bundle/binstubs" \
+  PATH=".bundle/binstubs:$PATH:/home/dependabot/.bundle/bin"
+
+COPY --from=build /home/dependabot/ .
