+++ docker/Gemfile
@@ -2,8 +2,6 @@
 
 source "https://rubygems.org"
 
-gem "rubocop"
-
 gem "dependabot-common", path: "../common"
 
 gemspec
